+++ lib/srv/app/server.go
@@ -31,6 +31,7 @@ import (
 	"github.com/gravitational/teleport"
 	apidefaults "github.com/gravitational/teleport/api/defaults"
 	"github.com/gravitational/teleport/api/types"
+	apiutils "github.com/gravitational/teleport/api/utils"
 	"github.com/gravitational/teleport/lib/auth"
 	"github.com/gravitational/teleport/lib/defaults"
 	"github.com/gravitational/teleport/lib/labels"
@@ -515,7 +516,7 @@ func (s *Server) getConfigForClient(info *tls.ClientHelloInfo) (*tls.Config, err
 
 	// Try and extract the name of the cluster that signed the client's certificate.
 	if info.ServerName != "" {
-		clusterName, err = auth.DecodeClusterName(info.ServerName)
+		clusterName, err = apiutils.DecodeClusterName(info.ServerName)
 		if err != nil {
 			if !trace.IsNotFound(err) {
 				s.log.Debugf("Ignoring unsupported cluster name %q.", info.ServerName)
